task12
Create a project with the Grid element that contains various controls: TextBoxes, Buttons, CheckBoxes, etc. Catch the moment when the Esc key is pressed on any of these controls without directly subscribing to events on each control. When the Esc key is pressed - disable all controls at once. There should be a way to enable them again.