visual element selector
Extend the project with the Snoop tool analogue with new functionality. Catch the mouse move action and check if the Ctrl+Shift combination is pressed. If it is, find a visual element under the mouse pointer and select it in your Snoop analogue's tree.

Note that you will need to store entire control instances for each tree item to be able to find and select a corresponding TreeViewItem. So, you need to think of how you store them and associate each control with each item.