Task 15 is finished
Create a simple View which shows a list of persons and allows editing each person's property.

For this, you need to create a Person object and fill ListBox with this object's instances. Than, override ListBox's ItemTemplate and place several elements to display all the object's properties in a row. After that, define several editors at the bottom of the View which allows you to edit a person's properties. ListBox should reflect each change as well.

Finally, use Triggers, DataTriggers or MultiTriggers in ItemTemplate to colorize items in ListBox based on a person's data and selection state. For example, colorize persons whose Age is greater than 50. Note that colors for selected and non-selected items should be different.