Snoop Data Templates added
Continue developing your analogue of the Snoop tool. Change the project so each TreeViewItem in the tree of elements contains a corresponding item from the visual tree. For this, you need to create a wrapper class and store elements from the visual tree in one of this object's properties. Create wrapper object instances for each item in the visual tree and store them in TreeViewItem's Header property.

To correctly display TreeViewItems, create a typed DataTemplate for the wrapper class and define elements in the manner so "rows" in TreeView are shown in the same manner as in the "Visual and Logical Tree - Snoop" and "Routed events - Snoop" tasks.