task refactored
1) Rewrite the "Traffic light" project according to the Custom Controls and OOP patterns.

2) Implement the CellStyle property which allows styling traffic light cells. Support specifying an image as the cell color. For example:

<local:Control.CellStyle>
  <Style TargetType="local:Cell">
    <Style.Triggers>
      <DataTrigger Binding="{Binding RelativeSource={RelativeSource Self}, Path=Color}" Value="Red">
        <Setter Property="Image" Value="ImageRed.png"/>
      </DataTrigger>
    </Style.Triggers>
  </Style>
</local:Control.CellStyle>

Here is a list of features to implement:

The traffic light should support a two- and three-color modes.
The traffic light should change lights automatically. It is necessary keep a current state and implement a way to set a delay before a certain light is switched.
Implement a custom event that is raised when the traffic light control switches to another state. This event's arguments should have information about the old and new state.
Add a button that allows switching lights manually.