task28 is finished
Creating a Behavior<T> class descendant is a way to implement a packed solution without creating control descendants. It implies that you move all logic to a Behavior<T> class descendant, which can be attached to a control in XAML/code. Please refer to this article for more information: http://www.wpftutorial.net/Behaviors.html

Our MVVM Framework (available as a part of our WPF subscription and as a free package in NuGet) contains its own implementation: see Behaviors