Custom control mvvm
Rewrite the traffic light control from the Custom Controls task so it fully meets the MVVM pattern.

According to this, you need to split the control functionality to the "logical" and "visual" parts.

The "logical" part should be implemented in View Model. For example, the logic to update states will be in the MainViewModel. CellViewModel will contain the State of the corresponding cell.

Here is a list of features to implement:

The traffic light should support a two- and three-color modes.
The traffic light should change lights automatically. It is necessary keep a current state and implement a way to set a delay before a certain light is switched.
Implement a custom event that is raised when the traffic light control switches to another state. This event's arguments should have information about the old and new state.
Add a button that allows switching lights manually.